Moved some files for clarity
