--- src/index.ts
@@ -1,15 +1,0 @@
-import * as analytics from './analytics';
-import * as donottrack from './donottrack';
-import * as history from './history';
-import * as SimpleAnalytics from './simpleanalytics';
-import * as storage from './storage';
-export { CoveoAnalyticsClient } from './analytics';
-export { CoveoUA, handleOneAnalyticsEvent } from './simpleanalytics';
-
-export {
-    analytics,
-    donottrack,
-    history,
-    SimpleAnalytics,
-    storage
-};
